refactor(phase4): isolate synthetic trust fixtures
